requirements.txt
manim
openai
rich
python-dotenv
